Added the Action class
Added the Action class to handle sending commands to the electronic device when a button is pushed.

The Button class has a member variable for an action but doesn't currently have a mechanism for calling peforming the action. I'm still deciding how best to do that.

I also added a spreadsheet in which I am starting to keep track of some of the IR codes.